requirements.txt
python-telegram-bot[job-queue]==21.5
aiohttp==3.9.5
anthropic==0.34.0
